move extract file
